Addition of fully working refactored unix_like Makefile and preparation for fully working refactored windows Makefile
