+++ test/src/test_argument_parser_parse_args.cpp
@@ -3,7 +3,7 @@
 
 #define AP_TESTING
 
-#include "../doctest.h"
+#include "doctest.h"
 
 #include <ap/argument_parser.hpp>
 #include <argument_parser_test_fixture.hpp>
