+++ test/src/test_true.cpp
@@ -1,4 +1,4 @@
-#include "../doctest.h"
+#include "doctest.h"
 
 TEST_CASE("true should be true") {
     REQUIRE(true);
